+++ aws_pricing/datadog_checks/aws_pricing/aws_pricing.py
@@ -1,110 +1,130 @@
 import json
-from collections import defaultdict
 
 import boto3
-from botocore.exceptions import ClientError
-from six import iteritems
-from six.moves import filter, map
 
 from datadog_checks.base import AgentCheck
-from datadog_checks.base.errors import CheckException
 
 
 class AwsPricingCheck(AgentCheck):
-    def check(self, instance):
+    # This will be the prefix of every metric and service check the integration sends
+    __NAMESPACE__ = "aws.pricing"
+
+    def __init__(self, name, init_config, instances):
+        super(AwsPricingCheck, self).__init__(name, init_config, instances)
+
+        self.region = self.instance.get("region_name") or "us-east-1"
+        self.services = self.instance.get("services") or []
+        self.filters = [
+            {"Field": filter["Key"], "Type": "TERM_MATCH", "Value": filter["Value"]}
+            for filter in self.instance.get("filters")
+        ]
+
+        if not self.services:
+            self.service_check(
+                "status",
+                self.WARNING,
+                message="No services defined; add at least one service in conf.yaml to collect pricing data.",
+            )
+
+        self.tag_fields = [
+            "enhancedNetworkingSupported",
+            "memory",
+            "vcpu",
+            "capacitystatus",
+            "instanceFamily",
+            "operatingSystem",
+            "regionCode",
+            "networkPerformance",
+            "instanceType",
+            "tenancy",
+            "usagetype",
+            "licenseModel",
+            "marketoption",
+            "termType",
+            "leaseContractLength",
+            "purchaseOption",
+            "unit",
+            "currency",
+        ]
+
+        self.client = boto3.client("pricing", region_name=self.region)
+
+    def check(self, _):
         try:
-            region_name = instance.get('region_name')
-            if not region_name:
-                region_name = 'us-east-1'
-
-            pricing_client = boto3.client('pricing', region_name=region_name)
-
-            service_codes = get_aws_service_codes(pricing_client)
-            rate_codes_dict = get_rate_codes_dict_from_instance(service_codes, instance)
-
-            # Python dictionaries evaluate to false when empty
-            if not rate_codes_dict:
-                message = 'No rate codes for existing AWS services were defined, please fix conf.yaml'
-                self.service_check('aws_pricing.status', self.CRITICAL, message=message)
-                raise CheckException(message)
-
-            missing_rate_codes = defaultdict(list)
-
-            for service_code, rate_codes in iteritems(rate_codes_dict):
-                for rate_code in rate_codes:
-                    price_dimensions = get_aws_prices(pricing_client, service_code, rate_code)
-
-                    if price_dimensions is None:
-                        missing_rate_codes[service_code].append(rate_code)
-                        continue
-
-                    name = 'aws.pricing.{}'.format(service_code.lower())
-                    price = get_price_from_price_dimensions(price_dimensions)
-                    tags = get_tags_from_price_dimensions(price_dimensions)
-
+            for service in self.services:
+                products = self.get_aws_products(service)
+                flattened_products = self.flatten_products(products)
+                for product in flattened_products:
+                    name = service.lower()
+                    price = float(product["price"])
+                    tags = [
+                        f'{self.lowercase(key)}:{value}'
+                        for key, value in product.items()
+                        if self.lowercase(key) in self.tag_fields
+                    ]
                     self.gauge(name, price, tags)
 
-            # Python dictionaries evaluate to true when not empty
-            if not missing_rate_codes:
-                self.service_check('aws_pricing.status', self.OK)
+            self.service_check("status", self.OK)
+        except Exception as e:
+            self.service_check("status", self.CRITICAL, message=str(e))
+
+    def lowercase(self, string):
+        return string[0].lower() + string[1:]
+
+    def flatten_products(self, raw_products):
+        products = []
+        for product in raw_products:
+            product_obj = json.loads(product)
+            products.extend(self.flatten_product(product_obj))
+        return products
+
+    def get_aws_products(self, service_code):
+        params = {
+            "ServiceCode": service_code,
+            "Filters": self.filters,
+            "FormatVersion": "aws_v1",
+        }
+
+        # Create an empty list to store the aggregated results
+        aggregated_results = []
+
+        while True:
+            response = self.client.get_products(**params)
+            aggregated_results.extend(response["PriceList"])
+
+            # Check if 'NextToken' exists in the response for pagination
+            if "NextToken" in response and response["NextToken"]:
+                params["NextToken"] = response["NextToken"]
             else:
-                message = 'Pricing data not found for these service rate codes: {}'.format(dict(missing_rate_codes))
-                self.service_check('aws_pricing.status', self.WARNING, message=message)
+                break
 
-        except ClientError as client_error:
-            self.service_check('aws_pricing.status', self.CRITICAL, message=str(client_error))
-            raise CheckException('Pricing Service client error: {}'.format(str(client_error)))
+        return aggregated_results
 
+    def flatten_product(self, product):
+        flattened_pricing_terms = []
 
-def get_rate_codes_dict_from_instance(service_codes, instance):
-    rate_codes_dict = {}
-    for service_code in service_codes:
-        instance_rate_codes = instance.get(service_code)
-        if instance_rate_codes is not None:
-            rate_codes_dict[service_code] = instance_rate_codes
+        # Loop through the "terms" which can be "OnDemand" or "Reserved"
+        for term_type, term_values in product["terms"].items():
+            # Loop through each SKU under "OnDemand" or "Reserved"
+            for sku, sku_values in term_values.items():
+                # Loop through each price dimension
+                for _, rate_values in sku_values["priceDimensions"].items():
+                    # Loop through each currency in pricePerUnit
+                    for currency, price in rate_values["pricePerUnit"].items():
+                        # Flatten attributes and rate_values into a single dictionary
+                        flattened_term = {
+                            **product["product"]["attributes"],  # Include all product attributes
+                            "termType": term_type,  # Include term type (OnDemand/Reserved)
+                            **sku_values.get("termAttributes", {}),  # Include any term attributes
+                            **rate_values,  # Include rate values like unit, endRange, etc.
+                            "currency": currency,  # Include the currency type
+                            "price": price,  # Include price for this currency
+                            "sku": sku,
+                            "effectiveDate": sku_values.get("effectiveDate", "N/A"),
+                            "offerTermCode": sku_values.get("offerTermCode", "N/A"),
+                        }
+                        # Remove the original pricePerUnit dictionary to prevent nested objects
+                        del flattened_term["pricePerUnit"]
 
-    return rate_codes_dict
-
-
-def get_aws_service_codes(pricing_client):
-    response = pricing_client.describe_services(FormatVersion='aws_v1')
-
-    service_codes = map(lambda service: service['ServiceCode'], response['Services'])
-
-    return service_codes
-
-
-def get_aws_prices(pricing_client, service_code, rate_code):
-    response = pricing_client.get_products(
-        FormatVersion='aws_v1',
-        ServiceCode=service_code,
-        Filters=[{'Type': 'TERM_MATCH', 'Field': 'RateCode', 'Value': rate_code}],
-        MaxResults=1,
-    )
-
-    price_dimensions = None
-
-    if len(response['PriceList']) > 0:
-        response_obj = json.loads(response['PriceList'][0])
-        terms = response_obj['terms'].values()
-        price_dimensions = find_price_dimensions_by_rate_code(rate_code, terms)
-
-    return price_dimensions
-
-
-def find_price_dimensions_by_rate_code(rate_code, terms):
-    rate_code_parts = rate_code.split('.')
-    term_code = '.'.join(rate_code_parts[:2])
-
-    term = next(filter(lambda term: term_code in term, terms))
-    price_dimensions = term[term_code]['priceDimensions'][rate_code]
-
-    return price_dimensions
-
-
-def get_tags_from_price_dimensions(price_dimensions):
-    return {'rate_code': price_dimensions['rateCode'], 'unit': price_dimensions['unit']}
-
-
-def get_price_from_price_dimensions(price_dimensions):
-    return float(price_dimensions['pricePerUnit']['USD'])
+                        flattened_pricing_terms.append(flattened_term)
+        return flattened_pricing_terms