+++ aws_pricing/CHANGELOG.md
@@ -1,2 +1,18 @@
-# CHANGELOG - Aws_pricing
+# CHANGELOG - AWS Pricing
 
+## 2.0.0 / 2023-10-20
+
+***Changed***:
+
+* New conf file structure to better align with the AWS Pricing API
+* Additional metadata tagged on the metrics
+* Updated to use the latest standards for Datadog checks
+
+***Added***:
+
+* Multiple products per check, no longer dependent on the rate code
+* Support for filters
+
+## 1.0.0 / 2019-07-18
+
+* Initial Release